src/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.0.1
discord==2.2.2
discord.py==2.2.2
frozenlist==1.3.3
idna==3.4
minio==7.1.14
multidict==6.0.4
numpy==1.24.3
pandas==2.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.28.2
six==1.16.0
structlog==23.1.0
tzdata==2023.3
Unidecode==1.3.6
urllib3==1.26.14
yarl==1.8.2
